create a native connection "class"
The native connection "class" hides the all the differences between the native
binding code and the javscript connection class. This way we can switch the
underlaying connection for a client easily and do not have to maintain different
query and pg "classes". Create a new Client like this:

new Client(config, ConnectionConstrutor)